+++ DbData/Get-DbSmo.ps1
@@ -49,10 +49,11 @@ function Get-DbSmo {
     [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSAvoidUsingPlainTextForPassword", "")]
     param (
         [Parameter(Mandatory, ValueFromPipeline, ValueFromPipelineByPropertyName, ParameterSetName = "SqlConnection", Position = 1)]
-        [System.Data.SqlClient.SqlConnection] $SqlConnection,
+        [Microsoft.Data.SqlClient.SqlConnection] $SqlConnection,
 
         [Parameter(Mandatory, ValueFromPipeline, ValueFromPipelineByPropertyName, ParameterSetName = "ServerInstance", Position = 1)]
         [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = "SqlConnection", Position = 2)]
+        [Alias("ServerName")]
         [Alias("SqlServerName")]
         [Alias("DataSource")]
         [string] $ServerInstance,